Errores menores corregidos.

+++ src/modelo/Usuario.java
@@ -127,7 +127,7 @@ public class Usuario {
      * @param cantidad - Importe de la venta.
      * @param concepto - Texto asociado a la venta.
      */
-    public void introducirVenta(int cantidad, String concepto) throws BeerBarException {
+    public void introducirVenta(float cantidad, String concepto) throws BeerBarException {
         Fecha hoy = new Fecha();
         hoy = hoy.getFechaActual();
         Transaccion transaccion = new Transaccion(cantidad, concepto, hoy, this);
@@ -141,7 +141,7 @@ public class Usuario {
      * @param cantidad - Importe del gasto.
      * @param concepto - Texto asociado al gasto.
      */
-    public void introducirGasto(int cantidad, String concepto) throws BeerBarException{
+    public void introducirGasto(float cantidad, String concepto) throws BeerBarException{
         Fecha hoy = new Fecha();
         hoy = hoy.getFechaActual();
         Transaccion transaccion = new Transaccion(-cantidad, concepto, hoy, this);
